rename to hercules
