+++ example_case_folders/example_sim_04/batch_script.sh
@@ -34,8 +34,8 @@ export HELICS_PORT=23405
 helics_broker -f 2 --consoleloglevel=trace --loglevel=debug   --local_port=$HELICS_PORT &
 
 
-# Need to set this to your emu_python folder
-# cd /home/pfleming/emu_python/emu_python
+# Need to set this to your hercules folder
+# cd /home/pfleming/hercules/hercules
 python3 emu_runscript.py emu_input_000.yaml >> logemu 2>&1  & # Start the controller center and pass in input file
 
 # Now go back to scratch folder and launch the job
