+++ example_case_folders/example_sim_04/emu_input_000.yaml
@@ -20,7 +20,7 @@ emu_comms:
   helics:
 
     config:
-        name: emu_python # What is the purpose of this name
+        name: hercules # What is the purpose of this name
         use_dash_frontend: False
         KAFKA: False
         KAFKA_topics: EMUV1py
