+++ example_case_folders/example_sim_06/batch_script.sh
@@ -29,8 +29,8 @@ export HELICS_PORT=32000
 # Set up the helics broker
 helics_broker -t zmq  -f 2 --loglevel="debug" --local_port=$HELICS_PORT & 
 
-# Need to set this to your emu_python folder
-# cd /home/pfleming/emu_python/emu_python
+# Need to set this to your hercules folder
+# cd /home/pfleming/hercules/hercules
 python3 emu_runscript.py emu_input_000.yaml >> logemu 2>&1  & # Start the controller center and pass in input file
 
 # Now go back to scratch folder and launch the job
