+++ example_case_folders/example_sim_07/batch_script.sh
@@ -28,8 +28,8 @@ module load netcdf-c/4.7.3/gcc-mpi
 # Set up the helics broker
 helics_broker -t zmq  -f 2 --loglevel="debug" & 
 
-# Need to set this to your emu_python folder
-# cd /home/pfleming/emu_python/emu_python
+# Need to set this to your hercules folder
+# cd /home/pfleming/hercules/hercules
 python3 emu_runscript.py emu_input_000.yaml >> logemu 2>&1  & # Start the controller center and pass in input file
 
 # Now go back to scratch folder and launch the job
