+++ example_case_folders/example_sim_08/bash_script.sh
@@ -7,8 +7,8 @@ conda activate emupy
 # Set up the helics broker
 helics_broker -t zmq  -f 2 --loglevel="debug" & 
 
-# Need to set this to your emu_python folder
-# cd /home/pfleming/emu_python/emu_python
+# Need to set this to your hercules folder
+# cd /home/pfleming/hercules/hercules
 python3 emu_runscript.py emu_input_000.yaml >> logemu 2>&1 & # Start the controller center and pass in input file
