+++ example_case_folders/example_sim_08/emu_runscript_dummy_amr.py
@@ -1,4 +1,4 @@
-from emu_python.dummy_amr_wind import launch_dummy_amr_wind
+from hercules.dummy_amr_wind import launch_dummy_amr_wind
 
 
 launch_dummy_amr_wind()
