+++ hercules/python_simulators/simple_solar.py
