+++ hercules/utilities.py
