+++ hercules/version.py
